+++ sql/queries/counts.sql
@@ -1,2 +1,2 @@
 -- name: GetSigns :many
-SELECT imageid::text, country_slug, state_slug, place_slug, county_slug FROM sign.vwhugohighwaysign;
+SELECT imageid::text, country_slug, state_slug, place_slug, county_slug, has_processed FROM sign.vwhugohighwaysign;